% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{db}
\alias{db}
\alias{db,JASPAR-method}
\alias{version,JASPAR-method}
\title{Access database from JASPAR object}
\usage{
db(object)

\S4method{db}{JASPAR}(object)

\S4method{version}{JASPAR}(object)
}
\arguments{
\item{object}{JASPAR class object}
}
\value{
Returns the path of the SQLite file in the AnnotationHub cache
}
\description{
The accessor function for retrieving the location of the
database location slot from the JASPAR object
}
\examples{

library(JASPAR)
jaspar <- JASPAR(version = 'JASPAR2024')
db(jaspar)

}
\author{
Damir Baranasic
}
\keyword{function}
