% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{JASPAR-class}
\alias{JASPAR-class}
\alias{JASPAR}
\title{JASPAR object class}
\usage{
JASPAR(version = "JASPAR2026")
}
\arguments{
\item{version}{The version of JASPAR database which should be loaded. 
Default is "JASPAR2026"}
}
\value{
JASPAR-class
}
\description{
The JASPAR object class is a thin class for storing the
path of JASPAR-style SQLite file.
}
\section{Slots}{

\describe{
\item{\code{db}}{Object of class \code{"character"} a character string of the path
of SQLite file.}

\item{\code{version}}{The version of JASPAR database which is loaded}
}}

\examples{

library(JASPAR)
library(RSQLite)

JASPAR <- JASPAR(version = 'JASPAR2026')
JASPARConnect <- RSQLite::dbConnect(RSQLite::SQLite(), db(JASPAR))
RSQLite::dbGetQuery(JASPARConnect, 'SELECT * FROM MATRIX LIMIT 5')
dbDisconnect(JASPARConnect)

}
\author{
Damir Baranasic
}
\keyword{classes}
