% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{version}
\alias{version}
\title{Access the database version from JASPAR object}
\usage{
version(object)
}
\arguments{
\item{object}{JASPAR class object}
}
\value{
Returns the version of the JASPAR satabase in the JASPAR object
}
\description{
The accessor function for retrieving the version of the
JASPAR database from the JASPAR object
}
\examples{

library(JASPAR)
JASPAR <- JASPAR()
version(JASPAR)

}
\author{
Damir Baranasic
}
\keyword{function}
